+++ tests/ErrorHandlerTest.php
@@ -1,12 +1,12 @@
-<?php 
+<?php
 /**
  * PHP library for handling exceptions and errors.
  * 
  * @author     Josantonius - info@josantonius.com
- * @copyright  Copyright (c) 2016 JST PHP Framework
+ * @copyright  Copyright (c) 2016-2017
  * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
- * @link       https://github.com/Josantonius/PHP-ErrorHandler
- * @since      File available since 1.0.0 - Update: 2017-02-14
+ * @link       https://github.com/Josantonius/ErrorHandler
+ * @since      File available since 1.0.0 - Update: 2017-02-21
  */
 
 namespace Josantonius\ErrorHandler\Tests;